requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community


# OpenAI Integration
langchain-openai
openai

# Gemini Integraion
langchain-google-genai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# UI using streamlit
streamlit


# youtube chatbot requirements
youtube-transcript-api
faiss-cpu
tiktoken

# Tools for Agent
pydantic 
duckduckgo-search 
langchain_experimental